Provided Files + testing
I did some testing to try to get an LED to turn on when i left or right click on the mouse. Quartus compiles and programs successfully yet nothing happens when I click. Unlikely to be due to pin assignments, most likely to be due to some sort of code error. I don't fully understand how mouse.vhd works so will need to do more research.