Simulating gravity in ModelSim
This functions as intended, I now just need to integrate this into bouncy ball. In this file, clk is analogous for the vertical sync and the click is for pd1 (mouse click). I need to change the range of the integer to save on memory (look into this). Change value on line 22 to change how often the speed increases.